requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pyodbc==5.1.0
python-dotenv==1.0.1
pydantic==2.6.3
msal==1.28.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0